+++ packages/backend/src/services/suggestion/suggestion.service.ts
@@ -1,5 +1,4 @@
 import {
-  FunctionParameters,
   getToolName,
   MessageRole,
   SUGGESTION_MODEL,
@@ -30,9 +29,7 @@ export const suggestionsResponseTool: OpenAI.Chat.Completions.ChatCompletionTool
       description:
         "Generate suggestions for the user based on the available components and context.",
       strict: true,
-      parameters: zodToJsonSchema(
-        SuggestionsResponseSchema,
-      ) as FunctionParameters,
+      parameters: zodToJsonSchema(SuggestionsResponseSchema),
     },
   };